Restructure, changed entity to user to match the documentation
