+++ src/Classes/EventSubs/Discount.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Eightfold\Eventbrite\Classes\EventSubs;
+
+use Eightfold\Eventbrite\Classes\Abstracts\EventSub;
+use Eightfold\Eventbrite\Interfaces\EventSubInterface;
+
+class Discount extends EventSub implements EventSubInterface
+{
+    protected function routeName()
+    {
+        return 'discounts';
+    }
+}
