+++ src/Interfaces/EventSubInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Eightfold\Eventbrite\Interfaces;
+
+interface EventSubInterface
+{
+    function routeName();
+}
